requirements.txt
python-telegram-bot==20.7
groq==0.8.0
APScheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
